Rename newnewznab provider file to newznab.py, ready for prime time I think
